# Upgrade Docusaurus - enable Docusaurus Faster