--- a/.changeset/small-ligers-doubt.md
+++ b/.changeset/small-ligers-doubt.md
@@ -1,0 +1,6 @@
+---
+"docusaurus-theme-redoc": minor
+"redocusaurus": patch
+---
+
+Upgrade Docusaurus - add support for Docusaurus Faster (RsPack)