--- a/packages/docusaurus-plugin-redoc/package.json
+++ b/packages/docusaurus-plugin-redoc/package.json
@@ -33,8 +33,8 @@
     "redoc": "2.1.5"
   },
   "devDependencies": {
-    "@docusaurus/types": "3.5.2-canary-6122",
-    "@docusaurus/utils": "3.5.2-canary-6122",
+    "@docusaurus/types": "3.5.2-canary-6125",
+    "@docusaurus/utils": "3.5.2-canary-6125",
     "core-js": "^3.36.0",
     "mobx": "^6.12.4",
     "react": "^18.3.1",