--- a/packages/docusaurus-theme-redoc/src/redocData.ts
+++ b/packages/docusaurus-theme-redoc/src/redocData.ts
@@ -1,7 +1,11 @@
 import type { RedocRawOptions } from 'redoc';
 import merge from 'lodash/merge';
 import { Config, loadConfig } from '@redocly/openapi-core';
-import type { GlobalData, RedocThemeOverrides, ThemeOptions } from './types/options';
+import type {
+  GlobalData,
+  RedocThemeOverrides,
+  ThemeOptions,
+} from './types/options';
 
 const defaultOptions: Partial<RedocRawOptions> = {
   scrollYOffset: 'nav.navbar',