--- a/packages/docusaurus-theme-redoc/src/theme/ApiDocMdx/ApiDocMdx.tsx
+++ b/packages/docusaurus-theme-redoc/src/theme/ApiDocMdx/ApiDocMdx.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo } from 'react';
 import Redoc from '@theme/Redoc';
 import useSpecData from '@theme/useSpecData';
-import { MdxProps as Props } from '../../types/common';
+import type { MdxProps as Props } from '../../types/common';
 import '../ApiSchema/styles.css';
 
 const ApiDocMdx: React.FC<Props> = ({ id }: Props): JSX.Element => {