--- a/packages/docusaurus-theme-redoc/src/theme/useSpecData.ts
+++ b/packages/docusaurus-theme-redoc/src/theme/useSpecData.ts
@@ -1,5 +1,5 @@
 import { useAllPluginInstancesData } from '@docusaurus/useGlobalData';
-import { SpecProps } from '../types/common';
+import type { SpecProps } from '../types/common';
 
 /**
  *