--- a/packages/docusaurus-theme-redoc/src/utils/useSpecOptions.ts
+++ b/packages/docusaurus-theme-redoc/src/utils/useSpecOptions.ts
@@ -7,8 +7,8 @@ import {
 import { useColorMode } from '@docusaurus/theme-common';
 import merge from 'lodash/merge';
 import '../global';
-import { RedocRawOptions } from 'redoc';
-import { SpecProps } from '../types/common';
+import type { RedocRawOptions } from 'redoc';
+import {type  SpecProps } from '../types/common';
 import { GlobalData } from '../types/options';
 
 /**