--- a/packages/redocusaurus/package.json
+++ b/packages/redocusaurus/package.json
@@ -52,8 +52,8 @@
     "node": ">=14"
   },
   "devDependencies": {
-    "@docusaurus/theme-common": "3.5.2-canary-6125",
-    "@docusaurus/utils": "3.5.2-canary-6125",
+    "@docusaurus/theme-common": "^3.6.0",
+    "@docusaurus/utils": "^3.6.0",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "react": "^18.3.1",