--- a/website/docusaurus.config.ts
+++ b/website/docusaurus.config.ts
@@ -10,6 +10,17 @@ if (process.env.VERCEL_URL) {
 }
 
 const config: Config = {
+  future: {
+    // experimental_faster: (process.env.DOCUSAURUS_FASTER ?? 'true') === 'true',
+    experimental_faster: {
+      rspackBundler: false,
+      swcJsLoader: false,
+      swcHtmlMinimizer: true,
+      swcJsMinimizer: true,
+      lightningCssMinimizer: true,
+      mdxCrossCompilerCache: true,
+    },
+  },
   title: 'Redocusaurus',
   tagline: 'OpenAPI solution for Docusaurus docs with Redoc',
   customFields: {