- Upgrade website to Docusaurus v3.6
- Make this plugin compatible with [Docusaurus Faster](https://github.com/facebook/docusaurus/issues/10556)
- Keep retro-compatibility with lower versions


## Benchmark

Cold builds - 3.3x faster:

```bash
hyperfine --prepare 'yarn workspace redocusaurus-website clear' --runs 3 'DOCUSAURUS_FASTER=false yarn build:website' 'DOCUSAURUS_FASTER=true yarn build:website'


Benchmark 1: DOCUSAURUS_FASTER=false yarn build:website
  Time (mean ± σ):     15.677 s ±  1.331 s    [User: 26.425 s, System: 5.220 s]
  Range (min … max):   14.732 s … 17.199 s    3 runs

Benchmark 2: DOCUSAURUS_FASTER=true yarn build:website
  Time (mean ± σ):      4.714 s ±  0.558 s    [User: 7.375 s, System: 1.040 s]
  Range (min … max):    4.185 s …  5.297 s    3 runs

Summary
  DOCUSAURUS_FASTER=true yarn build:website ran
    3.33 ± 0.48 times faster than DOCUSAURUS_FASTER=false yarn build:website
```

Warn builds - almost the same:

```bash
DOCUSAURUS_FASTER=false yarn build:website
hyperfine --runs 3 'DOCUSAURUS_FASTER=false yarn build:website' 'DOCUSAURUS_FASTER=true yarn build:website'

Benchmark 1: DOCUSAURUS_FASTER=false yarn build:website
  Time (mean ± σ):      4.637 s ±  0.308 s    [User: 5.878 s, System: 0.614 s]
  Range (min … max):    4.282 s …  4.842 s    3 runs

Benchmark 2: DOCUSAURUS_FASTER=true yarn build:website
  Time (mean ± σ):      4.230 s ±  0.139 s    [User: 6.986 s, System: 1.055 s]
  Range (min … max):    4.075 s …  4.346 s    3 runs

Summary
  DOCUSAURUS_FASTER=true yarn build:website ran
    1.10 ± 0.08 times faster than DOCUSAURUS_FASTER=false yarn build:website
```


## Tests

Note: I'd recommend testing in particular Prism code blocks because I had weird SSG errors that I fixed with https://github.com/facebook/docusaurus/pull/10618