@wilhelmer 2024-11-08
Thank you for this! Would love to see this released in the near future 👌